% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_groups.R
\name{count_groups}
\alias{count_groups}
\title{Count the number of groups in a predictor variable}
\usage{
count_groups(data, predictor)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{predictor}{A string giving the name of the predictor variable.}
}
\value{
An integer giving the number of unique groups in the predictor.
}
\description{
Determines how many unique groups (levels or categories) the predictor
variable contains. This function is designed to support the decision tree
for choosing the appropriate statistical test.
}
\examples{
count_groups(PlantGrowth, predictor = "group")

}
