% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_test.R
\name{run_test}
\alias{run_test}
\title{Run a statistical test based on variable types and group structure}
\usage{
run_test(data, outcome, predictor)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome}{A string giving the name of the outcome variable.}

\item{predictor}{A string giving the name of the predictor variable.}
}
\value{
A list containing:
\describe{
  \item{test}{The name of the test performed}
  \item{result}{The model or test output object}
}
}
\description{
This function selects and runs an appropriate statistical test based on the
outcome and predictor variables. It uses \code{choose_test()} to determine
which test to run, and then performs the corresponding analysis.
}
\examples{
run_test(PlantGrowth, outcome = "weight", predictor = "group")

}
