% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_test.R
\name{choose_test}
\alias{choose_test}
\title{Choose an appropriate statistical test}
\usage{
choose_test(data, outcome, predictor)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome}{A string giving the name of the outcome variable.}

\item{predictor}{A string giving the name of the predictor variable.}
}
\value{
A character string naming the recommended statistical test.
}
\description{
Based on the variable types (numeric or categorical) and the number of
groups in the predictor, this function selects an appropriate statistical
test. It uses helper functions \code{identify_var_type()} and
\code{count_groups()} to determine the structure of the data.
}
\examples{
choose_test(PlantGrowth, outcome = "weight", predictor = "group")

}
