% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_groups.R
\name{count_groups}
\alias{count_groups}
\title{Count the number of groups in a predictor variable}
\usage{
count_groups(x)
}
\arguments{
\item{x}{A vector (factor or character).}
}
\value{
Integer number of unique groups.
}
\description{
Determines how many unique groups (levels or categories) the predictor
variable contains. This function is designed to support the decision tree
for choosing the appropriate statistical test.
}
