% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_var_type.R
\name{identify_var_type}
\alias{identify_var_type}
\title{Identify variable type}
\usage{
identify_var_type(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A character string: "numeric" or "categorical".
}
\description{
Determines whether a variable is numeric or categorical.
}
\examples{
identify_var_type(1:10)
identify_var_type(factor(c("A", "B")))

}
